+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Tests for "Defibrillators".
+- Tests for "Horse-racing duals".
 
 ## [1.0.0] - 2022-02-26
 ### Added